reduce modCount type for local value

+++ modules/base/execution-api/src/main/java/consulo/execution/internal/ExecutionActionValue.java
@@ -66,7 +66,7 @@ public class ExecutionActionValue implements LocalizeValueWithMnemonic {
     }
 
     @Override
-    public long getModificationCount() {
+    public byte getModificationCount() {
         return myOriginal.getModificationCount();
     }
